Enable eglGetPlatformDisplay() for 1.5
eglGetPlatformDisplay() may be used to get a display for a rendering
context. However, epoxy_conservative_egl_version() cannot report a
correct version without a display and "conservatively" reports 1.4
instead. This disables eglGetPlatformDisplay(), which is available only
since 1.5.

To fix this issue, query the client version for eglGetPlatformDisplay().
The client version is a concept introduced with 1.5, and does not
require a display to query. Failing the query implies the version is
older than 1.5 and eglGetPlatformDisplay() is not supported.