Disable desktop-only extension functions for ES
glDebugMessageCallback() is defined with GL_KHR_debug and on OpenGL ES
3.2, but it has the KHR suffix on the older OpenGL ES versions with
GL_KHR_debug. For OpenGL ES, libepoxy needs to resolve the function
with the KHR suffix if GL_KHR_debug is present, and to resolve it
without the suffix otherwise.

However, libepoxy incorrectly tries to resolve the function without
the suffix if GL_KHR_debug is present because it is confused with the
extension definition for desktop that doesn't use the suffix.

This can result in a non-functional resolution of
glDebugMessageCallback(). For example, an OpenGL ES implementation,
ANGLE can always resolve the function whether the name is suffixed or
not. However, the function resolved without the suffix is non-functional
if it exposes a version older than 3.2.

To fix this issue, distinguish between the extension definitions for
desktop and those for OpenGL ES. This prevents libepoxy to try resolving
glDebugMessageCallback() without the KHR suffix on OpenGL ES and lets it
try resolving it with the suffix when GL_KHR_debug is present.